feat(ssr): split files to implement ssr
